+++ iam-policy-autopilot-policy-generation/src/lsp/mod.rs
@@ -8,11 +8,13 @@
 mod error;
 #[doc(hidden)]
 pub mod gopls;
+mod ty;
 
 #[cfg(any(test, feature = "integ-test"))]
 pub mod test_utils;
 
 pub use error::LspError;
+pub use ty::TyLspClient;
 
 use std::collections::HashSet;
 use std::ops::ControlFlow;
@@ -54,23 +56,6 @@ pub trait LspServerConfig {
     }
 }
 
-/// Configuration for the ty Python type checker.
-pub struct TyConfig;
-
-impl LspServerConfig for TyConfig {
-    fn binary_name(&self) -> &'static str {
-        "ty"
-    }
-
-    fn args(&self) -> &[&str] {
-        &["server"]
-    }
-
-    fn language_id(&self) -> &'static str {
-        "python"
-    }
-}
-
 /// Options for configuring `LspClient` behavior.
 #[derive(Debug)]
 pub struct LspClientOptions {
@@ -78,8 +63,9 @@ pub struct LspClientOptions {
     pub open_document_timeout: Duration,
     /// Timeout for the initialize handshake.
     pub initialize_timeout: Duration,
-    /// Timeout for hover requests.
-    pub hover_timeout: Duration,
+    /// Timeout for individual server requests (hover, document symbols, call
+    /// hierarchy).
+    pub request_timeout: Duration,
     /// Timeout for shutdown.
     pub shutdown_timeout: Duration,
 }
@@ -89,7 +75,7 @@ impl Default for LspClientOptions {
         Self {
             open_document_timeout: Duration::from_secs(1),
             initialize_timeout: Duration::from_secs(10),
-            hover_timeout: Duration::from_secs(5),
+            request_timeout: Duration::from_secs(5),
             shutdown_timeout: Duration::from_secs(2),
         }
     }
@@ -309,9 +295,9 @@ impl<C: LspServerConfig> LspClient<C> {
             work_done_progress_params: WorkDoneProgressParams::default(),
         };
 
-        let response = timeout(self.options.hover_timeout, self.server.hover(params))
+        let response = timeout(self.options.request_timeout, self.server.hover(params))
             .await
-            .map_err(|_| LspError::Timeout(self.options.hover_timeout))?
+            .map_err(|_| LspError::Timeout(self.options.request_timeout))?
             .map_err(|e| LspError::ServerError(format!("{e}")))?;
 
         Ok(response.and_then(|hover| extract_type_from_hover(&hover)))
@@ -329,11 +315,11 @@ impl<C: LspServerConfig> LspClient<C> {
         };
 
         let response = timeout(
-            self.options.hover_timeout,
+            self.options.request_timeout,
             self.server.document_symbol(params),
         )
         .await
-        .map_err(|_| LspError::Timeout(self.options.hover_timeout))?
+        .map_err(|_| LspError::Timeout(self.options.request_timeout))?
         .map_err(|e| LspError::ServerError(format!("{e}")))?;
 
         Ok(response)
@@ -358,11 +344,11 @@ impl<C: LspServerConfig> LspClient<C> {
         };
 
         let response = timeout(
-            self.options.hover_timeout,
+            self.options.request_timeout,
             self.server.prepare_call_hierarchy(params),
         )
         .await
-        .map_err(|_| LspError::Timeout(self.options.hover_timeout))?
+        .map_err(|_| LspError::Timeout(self.options.request_timeout))?
         .map_err(|e| LspError::ServerError(format!("{e}")))?;
 
         Ok(response)
@@ -382,11 +368,11 @@ impl<C: LspServerConfig> LspClient<C> {
         };
 
         let response = timeout(
-            self.options.hover_timeout,
+            self.options.request_timeout,
             self.server.outgoing_calls(params),
         )
         .await
-        .map_err(|_| LspError::Timeout(self.options.hover_timeout))?
+        .map_err(|_| LspError::Timeout(self.options.request_timeout))?
         .map_err(|e| LspError::ServerError(format!("{e}")))?;
 
         Ok(response)
@@ -459,24 +445,6 @@ fn non_empty(s: &str) -> Option<String> {
         None
     } else {
         Some(s.to_string())
-    }
-}
-
-/// Convenience type alias for the ty Python type checker client.
-pub type TyLspClient = LspClient<TyConfig>;
-
-impl TyLspClient {
-    /// Create a new ty LSP client with default options.
-    pub async fn create(workspace_root: impl AsRef<Path>) -> Result<Self, LspError> {
-        Self::new(TyConfig, workspace_root).await
-    }
-
-    /// Create a new ty LSP client with custom options.
-    pub async fn create_with_options(
-        workspace_root: impl AsRef<Path>,
-        options: LspClientOptions,
-    ) -> Result<Self, LspError> {
-        Self::with_options(TyConfig, workspace_root, options).await
     }
 }
